Handle deeplink for offers
