+++ app/src/main/java/in/testpress/testpress/ui/SplashScreenActivity.java
@@ -27,6 +27,7 @@ import in.testpress.testpress.authenticator.LoginActivity;
 import in.testpress.testpress.authenticator.ResetPasswordActivity;
 import in.testpress.testpress.core.Constants;
 import in.testpress.testpress.core.TestpressService;
+import in.testpress.testpress.ui.utils.DeeplinkHandler;
 import in.testpress.testpress.util.CommonUtils;
 import in.testpress.testpress.util.UpdateAppDialogManager;
 import in.testpress.util.Assert;
@@ -60,185 +61,19 @@ public class SplashScreenActivity extends Activity {
         Injector.inject(this);
         ButterKnife.inject(this);
         UpdateAppDialogManager.monitor(this);
+        final DeeplinkHandler deeplinkHandler = new DeeplinkHandler(this, serviceProvider);
         new Handler().postDelayed(new Runnable() {
 
             @Override
             public void run() {
-                Uri uri = getIntent().getData();
-                if (uri != null && uri.getPathSegments().size() > 0) {
-                    List<String> pathSegments = uri.getPathSegments();
-                    switch (pathSegments.get(0)) {
-                        case "p":
-                            Intent intent = new Intent(SplashScreenActivity.this, PostActivity.class);
-                            intent.putExtra("shortWebUrl", uri.toString());
-                            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK |
-                                    Intent.FLAG_ACTIVITY_CLEAR_TASK);
-                            intent.putExtra(Constants.IS_DEEP_LINK, true);
-                            startActivity(intent);
-                            finish();
-                            break;
-                        case "posts":
-                            Intent postsIntent =
-                                    new Intent(SplashScreenActivity.this, PostsListActivity.class);
-                            postsIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK |
-                                    Intent.FLAG_ACTIVITY_CLEAR_TASK);
-
-                            postsIntent.putExtra(Constants.IS_DEEP_LINK, true);
-                            startActivity(postsIntent);
-                            finish();
-                            break;
-                        case "exams":
-                            authenticateUser(uri);
-                            break;
-                        case "user":
-                        case "profile":
-                            if (pathSegments.size() == 1) {
-                                gotoActivity(ProfileDetailsActivity.class, true);
-                            } else {
-                                CommonUtils.openUrlInBrowser(SplashScreenActivity.this, uri);
-                                finish();
-                            }
-                            break;
-                        case "password":
-                            gotoActivity(ResetPasswordActivity.class, false);
-                            break;
-                        case "analytics":
-                            authenticateUser(uri);
-                            break;
-                        case "documents":
-                            gotoActivity(DocumentsListActivity.class, true);
-                            break;
-                        case "login":
-                            gotoActivity(MainActivity.class, true);
-                            break;
-                        case "activate":
-                            gotoAccountActivate(uri.getPath());
-                            break;
-                        case "chapters":
-                            authenticateUser(uri);
-                            break;
-                        case "courses":
-                        case "learn":
-                        case "leaderboard":
-                        case "dashboard":
-                            gotoHome();
-                            break;
-                        case "store":
-                        case "market":
-                        case "products":
-                            authenticateUser(uri);
-                            break;
-                        default:
-                            CommonUtils.openUrlInBrowser(SplashScreenActivity.this, uri);
-                            finish();
-                            break;
-                    }
-                } else {
-                    // This method will be executed once the timer is over
-                    // Start app main activity
-                    gotoHome();
-                }
+                deeplinkHandler.handleDeepLinkUrl(getIntent().getData());
             }
         }, SPLASH_TIME_OUT);
     }
 
+
     private void gotoHome() {
         Intent intent = new Intent(this, MainActivity.class);
-        startActivity(intent);
-        finish();
-    }
-
-    private void authenticateUser(final Uri uri) {
-        final Activity activity = SplashScreenActivity.this;
-        final List<String> pathSegments = uri.getPathSegments();
-        CommonUtils.getAuth(activity, serviceProvider,
-                new CommonUtils.CheckAuthCallBack() {
-                    @Override
-                    public void onSuccess(TestpressService testpressService) {
-                        TestpressSession testpressSession = TestpressSdk.getTestpressSession(activity);
-                        Assert.assertNotNull("TestpressSession must not be null.", testpressSession);
-                        switch (pathSegments.get(0)) {
-                            case "exams":
-                                if (pathSegments.size() == 2) {
-                                    if (!pathSegments.get(1).equals("available") ||
-                                            !pathSegments.get(1).equals("upcoming") ||
-                                            !pathSegments.get(1).equals("history")) {
-
-                                        // If exam slug is present, directly goto the start exam screen
-                                        TestpressExam.showExamAttemptedState(
-                                                activity,
-                                                pathSegments.get(1),
-                                                testpressSession
-                                        );
-                                        return;
-                                    }
-                                }
-                                // Show exams list
-                                TestpressExam.show(activity, testpressSession);
-                                finish();
-                                break;
-
-                            case "analytics":
-                                TestpressExam.showAnalytics(activity, SUBJECT_ANALYTICS_PATH,
-                                        testpressSession);
-                                break;
-                            case "chapters":
-                                deepLinkToChapter(uri, testpressSession);
-                                break;
-                            case "products":
-                                deepLinkToProduct(uri, testpressSession);
-                                break;
-                        }
-                    }
-                });
-    }
-
-    private void deepLinkToChapter(Uri uri, TestpressSession testpressSession) {
-        final List<String> pathSegments = uri.getPathSegments();
-        switch (pathSegments.size()) {
-            case 1:
-                // /chapters/
-                gotoHome();
-                break;
-            case 2:
-                // Contents list url - /chapters/chapter-slug/
-                String chapterAPI = BASE_URL + CHAPTERS_PATH + uri.getLastPathSegment() + "/";
-                TestpressCourse.showChapterContents(this, chapterAPI, testpressSession);
-                break;
-            case 3:
-                // Content detail url - /chapters/chapter-slug/{content-id}/
-                TestpressCourse.showContentDetail(this, uri.getLastPathSegment(), testpressSession);
-                break;
-        }
-    }
-
-    private void deepLinkToProduct(Uri uri, TestpressSession testpressSession) {
-        final List<String> pathSegments = uri.getPathSegments();
-        if (pathSegments.size() == 2) {
-            // Product detail url - /products/product-slug/
-            String productSlug = uri.getLastPathSegment();
-            assert productSlug != null;
-            TestpressStore.showProduct(this, productSlug, testpressSession);
-        } else {
-            TestpressStore.show(this, testpressSession);
-        }
-    }
-
-    private void gotoAccountActivate(String activateUrlFrag) {
-        Intent intent = new Intent(SplashScreenActivity.this, AccountActivateActivity.class);
-        intent.putExtra(AccountActivateActivity.ACTIVATE_URL_FRAG, activateUrlFrag);
-        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-        startActivity(intent);
-        finish();
-    }
-
-    private void gotoActivity(Class activityClass, boolean requireAuthentication) {
-        if (requireAuthentication && !CommonUtils.isUserAuthenticated(SplashScreenActivity.this)) {
-            activityClass = LoginActivity.class;
-        }
-        Intent intent = new Intent(SplashScreenActivity.this, activityClass);
-        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-        intent.putExtra(Constants.IS_DEEP_LINK, true);
         startActivity(intent);
         finish();
     }
