+++ app/src/main/java/in/testpress/testpress/ui/utils/DeeplinkHandler.java
@@ -1,0 +1,211 @@
+package in.testpress.testpress.ui.utils;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.net.Uri;
+
+import java.util.List;
+
+import in.testpress.core.TestpressSdk;
+import in.testpress.core.TestpressSession;
+import in.testpress.course.TestpressCourse;
+import in.testpress.exam.TestpressExam;
+import in.testpress.store.TestpressStore;
+import in.testpress.testpress.TestpressServiceProvider;
+import in.testpress.testpress.authenticator.LoginActivity;
+import in.testpress.testpress.authenticator.ResetPasswordActivity;
+import in.testpress.testpress.core.Constants;
+import in.testpress.testpress.core.TestpressService;
+import in.testpress.testpress.ui.AccountActivateActivity;
+import in.testpress.testpress.ui.DocumentsListActivity;
+import in.testpress.testpress.ui.MainActivity;
+import in.testpress.testpress.ui.PostActivity;
+import in.testpress.testpress.ui.PostsListActivity;
+import in.testpress.testpress.ui.ProfileDetailsActivity;
+import in.testpress.testpress.ui.SplashScreenActivity;
+import in.testpress.testpress.util.CommonUtils;
+import in.testpress.util.Assert;
+
+import static in.testpress.exam.network.TestpressExamApiClient.SUBJECT_ANALYTICS_PATH;
+import static in.testpress.testpress.BuildConfig.BASE_URL;
+import static in.testpress.testpress.core.Constants.Http.CHAPTERS_PATH;
+
+public class DeeplinkHandler {
+    private Activity activity;
+    TestpressServiceProvider serviceProvider;
+
+    public DeeplinkHandler(Activity activity, TestpressServiceProvider serviceProvider) {
+        this.activity = activity;
+        this.serviceProvider = serviceProvider;
+    }
+
+    public void handleDeepLinkUrl(Uri uri) {
+        if (uri != null && uri.getPathSegments().size() > 0) {
+            List<String> pathSegments = uri.getPathSegments();
+            switch (pathSegments.get(0)) {
+                case "p":
+                    Intent intent = new Intent(activity, PostActivity.class);
+                    intent.putExtra("shortWebUrl", uri.toString());
+                    intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK |
+                            Intent.FLAG_ACTIVITY_CLEAR_TASK);
+                    intent.putExtra(Constants.IS_DEEP_LINK, true);
+                    activity.startActivity(intent);
+                    activity.finish();
+                    break;
+                case "posts":
+                    Intent postsIntent =
+                            new Intent(activity, PostsListActivity.class);
+                    postsIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK |
+                            Intent.FLAG_ACTIVITY_CLEAR_TASK);
+
+                    postsIntent.putExtra(Constants.IS_DEEP_LINK, true);
+                    activity.startActivity(postsIntent);
+                    activity.finish();
+                    break;
+                case "exams":
+                    authenticateUser(uri);
+                    break;
+                case "user":
+                case "profile":
+                    if (pathSegments.size() == 1) {
+                        gotoActivity(ProfileDetailsActivity.class, true);
+                    } else {
+                        CommonUtils.openUrlInBrowser(activity, uri);
+                        activity.finish();
+                    }
+                    break;
+                case "password":
+                    gotoActivity(ResetPasswordActivity.class, false);
+                    break;
+                case "analytics":
+                    authenticateUser(uri);
+                    break;
+                case "documents":
+                    gotoActivity(DocumentsListActivity.class, true);
+                    break;
+                case "login":
+                    gotoActivity(MainActivity.class, true);
+                    break;
+                case "activate":
+                    gotoAccountActivate(uri.getPath());
+                    break;
+                case "chapters":
+                    authenticateUser(uri);
+                    break;
+                case "courses":
+                case "learn":
+                case "leaderboard":
+                case "dashboard":
+                    gotoHome();
+                    break;
+                case "store":
+                case "market":
+                case "products":
+                    authenticateUser(uri);
+                    break;
+                default:
+                    CommonUtils.openUrlInBrowser(activity, uri);
+                    activity.finish();
+                    break;
+            }
+        } else {
+            gotoHome();
+        }
+    }
+
+    private void authenticateUser(final Uri uri) {
+        final List<String> pathSegments = uri.getPathSegments();
+        CommonUtils.getAuth(activity, serviceProvider,
+                new CommonUtils.CheckAuthCallBack() {
+                    @Override
+                    public void onSuccess(TestpressService testpressService) {
+                        TestpressSession testpressSession = TestpressSdk.getTestpressSession(activity);
+                        Assert.assertNotNull("TestpressSession must not be null.", testpressSession);
+                        switch (pathSegments.get(0)) {
+                            case "exams":
+                                if (pathSegments.size() == 2) {
+                                    if (!pathSegments.get(1).equals("available") ||
+                                            !pathSegments.get(1).equals("upcoming") ||
+                                            !pathSegments.get(1).equals("history")) {
+
+                                        // If exam slug is present, directly goto the start exam screen
+                                        TestpressExam.showExamAttemptedState(
+                                                activity,
+                                                pathSegments.get(1),
+                                                testpressSession
+                                        );
+                                        return;
+                                    }
+                                }
+                                // Show exams list
+                                TestpressExam.show(activity, testpressSession);
+                                activity.finish();
+                                break;
+
+                            case "analytics":
+                                TestpressExam.showAnalytics(activity, SUBJECT_ANALYTICS_PATH,
+                                        testpressSession);
+                                break;
+                            case "chapters":
+                                deepLinkToChapter(uri, testpressSession);
+                                break;
+                            case "products":
+                                deepLinkToProduct(uri, testpressSession);
+                                break;
+                        }
+                    }
+                });
+    }
+
+    private void deepLinkToChapter(Uri uri, TestpressSession testpressSession) {
+        final List<String> pathSegments = uri.getPathSegments();
+        switch (pathSegments.size()) {
+            case 1:
+                gotoHome();
+                break;
+            case 2:
+                String chapterAPI = BASE_URL + CHAPTERS_PATH + uri.getLastPathSegment() + "/";
+                TestpressCourse.showChapterContents(activity, chapterAPI, testpressSession);
+                break;
+            case 3:
+                TestpressCourse.showContentDetail(activity, uri.getLastPathSegment(), testpressSession);
+                break;
+        }
+    }
+
+    private void deepLinkToProduct(Uri uri, TestpressSession testpressSession) {
+        final List<String> pathSegments = uri.getPathSegments();
+        if (pathSegments.size() == 2) {
+            String productSlug = uri.getLastPathSegment();
+            assert productSlug != null;
+            TestpressStore.showProduct(activity, productSlug, testpressSession);
+        } else {
+            TestpressStore.show(activity, testpressSession);
+        }
+    }
+
+    private void gotoAccountActivate(String activateUrlFrag) {
+        Intent intent = new Intent(activity, AccountActivateActivity.class);
+        intent.putExtra(AccountActivateActivity.ACTIVATE_URL_FRAG, activateUrlFrag);
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+        activity.startActivity(intent);
+        activity.finish();
+    }
+
+    private void gotoActivity(Class activityClass, boolean requireAuthentication) {
+        if (requireAuthentication && !CommonUtils.isUserAuthenticated(activity)) {
+            activityClass = LoginActivity.class;
+        }
+        Intent intent = new Intent(activity, activityClass);
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+        intent.putExtra(Constants.IS_DEEP_LINK, true);
+        activity.startActivity(intent);
+        activity.finish();
+    }
+
+    private void gotoHome() {
+        Intent intent = new Intent(activity, MainActivity.class);
+        activity.startActivity(intent);
+        activity.finish();
+    }
+}
